+++ XayyatMario/src/main/java/com/mammedbrk/model/component/pipe/TelePipe.java
@@ -8,7 +8,7 @@ public abstract class TelePipe extends Pipe {
     public TelePipe() {
     }
 
-    public TelePipe(int x, int y, Section section) {
+    public TelePipe(double x, double y, Section section) {
         super(x, y);
         this.section = section;
     }
